+++ includes/ft_printf_structs.h
@@ -41,7 +41,7 @@ typedef struct	s_params
 	size_t		n;
 	t_flags		*flags;
 	size_t		width;
-	size_t		precision;
+	intmax_t	precision;
 	t_length	length;
 	t_output	*output;
 	t_output	*prefix;
